AI-todo/phase-5/requirements.txt
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
argon2-cffi>=23.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
slowapi>=0.1.9
openai>=1.0.0
mcp>=1.0.0
aiokafka>=0.10.0
kafka-python>=2.0.2
python-dateutil>=2.8.2